Default dtype to float64 via _constants; propagate dtype through predictors
Manifold math generally needs float64 (the ambient cosh/sin coordinates overflow
float32 at high |K|*sigma^2*dim), so make it the library-wide default.

- Add manify/_constants.py with DEFAULT_DTYPE (float64) and DEFAULT_DEVICE.
- Manifold/ProductManifold default to these constants; ProductManifold gains full
  dtype support (factors, projection matrix, sampling, log-likelihood, stereographic
  conversions, gaussian_mixture) so it no longer mixes float32/float64 internally.
- Propagate the manifold dtype through the ML components that previously assumed
  float32: decision-tree scatter buffers follow their source dtype; KappaGCN/
  Stereographic layers cast parameters to the manifold dtype; SVM stores/queries in
  pm.dtype; KappaGCN/transformer cast labels to pm.dtype for the loss.

Epsilons (learning rates, tolerances, jitters, numeric guards) are intentionally left
as documented local literals -- they are heterogeneous and not a shared constant.

Tests: default-dtype test; float32/float64 propagation; quantile/label/regression
paths updated for the float64 default. Full suite green except pre-existing
network-dependent dataset-download tests.

+++ manify/predictors/kappa_gcn.py
@@ -202,11 +202,11 @@ class KappaGCN(BasePredictor, torch.nn.Module):
             y = y.long()
         elif self.task == "regression":
             loss_fn = torch.nn.MSELoss()
-            y = y.float()
+            y = y.to(self.pm.dtype)
         elif self.task == "link_prediction":
             loss_fn = torch.nn.BCEWithLogitsLoss()
             # y = y.flatten().float()
-            y = y.float()
+            y = y.to(self.pm.dtype)
         else:
             raise ValueError("Invalid task!")
